feat: move scrape into package
